moved indexable-builder-versions to values folder
