+++ src/builders/indexable-term-builder.php
@@ -2,7 +2,7 @@
 
 namespace Yoast\WP\SEO\Builders;
 
-use Yoast\WP\SEO\Config\Indexable_Builder_Versions;
+use Yoast\WP\SEO\Values\Indexables\Indexable_Builder_Versions;
 use Yoast\WP\SEO\Exceptions\Indexable\Invalid_Term_Exception;
 use Yoast\WP\SEO\Exceptions\Indexable\Term_Not_Found_Exception;
 use Yoast\WP\SEO\Helpers\Taxonomy_Helper;
