+++ src/services/indexables/indexable-version-manager.php
@@ -2,7 +2,7 @@
 
 namespace Yoast\WP\SEO\Services\Indexables;
 
-use Yoast\WP\SEO\Config\Indexable_Builder_Versions;
+use Yoast\WP\SEO\Values\Indexables\Indexable_Builder_Versions;
 use Yoast\WP\SEO\Models\Indexable;
 
 /**
