+++ tests/unit/config/indexable-builder-versions-test.php
@@ -2,7 +2,7 @@
 
 namespace Yoast\WP\SEO\Tests\Unit\Config;
 
-use Yoast\WP\SEO\Config\Indexable_Builder_Versions;
+use Yoast\WP\SEO\Values\Indexables\Indexable_Builder_Versions;
 use Yoast\WP\SEO\Tests\Unit\Doubles\Config\Indexable_Builder_Versions_Double;
 use Yoast\WP\SEO\Tests\Unit\TestCase;
